Add Guide section: Testing Expressions
