+++ mathics/builtin/arithmetic.py
@@ -48,7 +48,6 @@ from mathics.core.attributes import (
 )
 from mathics.core.convert.expression import to_expression
 from mathics.core.convert.mpmath import from_mpmath
-from mathics.core.convert.python import from_python
 from mathics.core.convert.sympy import SympyExpression, from_sympy, sympy_symbol_prefix
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
@@ -68,13 +67,11 @@ from mathics.core.systemsymbols import (
     SymbolAnd,
     SymbolComplexInfinity,
     SymbolDirectedInfinity,
-    SymbolExpandAll,
     SymbolIndeterminate,
     SymbolInfix,
     SymbolOverflow,
     SymbolPiecewise,
     SymbolPossibleZeroQ,
-    SymbolSimplify,
     SymbolTable,
     SymbolUndefined,
 )
@@ -798,29 +795,6 @@ class Integer_(Builtin):
     name = "Integer"
 
 
-class NumberQ(Test):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/NumberQ.html</url>
-
-    <dl>
-      <dt>'NumberQ[$expr$]'
-      <dd>returns 'True' if $expr$ is an explicit number, and 'False' otherwise.
-    </dl>
-
-    >> NumberQ[3+I]
-     = True
-    >> NumberQ[5!]
-     = True
-    >> NumberQ[Pi]
-     = False
-    """
-
-    summary_text = "test whether an expression is a number"
-
-    def test(self, expr):
-        return isinstance(expr, Number)
-
-
 class Piecewise(SympyFunction):
     """
     <url>:WMA link:https://reference.wolfram.com/language/ref/Piecewise.html</url>
@@ -914,75 +888,6 @@ class Piecewise(SympyFunction):
         if str(args[-1].elements[1]).startswith("System`_True__Dummy_"):
             args[-1].elements[1] = SymbolTrue
         return Expression(self.get_name(), args)
-
-
-class PossibleZeroQ(SympyFunction):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/PossibleZeroQ.html</url>
-
-    <dl>
-      <dt>'PossibleZeroQ[$expr$]'
-      <dd>returns 'True' if basic symbolic and numerical methods suggest that expr has value zero, and 'False' otherwise.
-    </dl>
-
-    Test whether a numeric expression is zero:
-    >> PossibleZeroQ[E^(I Pi/4) - (-1)^(1/4)]
-     = True
-
-    The determination is approximate.
-
-    Test whether a symbolic expression is likely to be identically zero:
-    >> PossibleZeroQ[(x + 1) (x - 1) - x^2 + 1]
-     = True
-
-
-    >> PossibleZeroQ[(E + Pi)^2 - E^2 - Pi^2 - 2 E Pi]
-     = True
-
-    Show that a numeric expression is nonzero:
-    >> PossibleZeroQ[E^Pi - Pi^E]
-     = False
-
-    >> PossibleZeroQ[1/x + 1/y - (x + y)/(x y)]
-     = True
-
-    Decide that a numeric expression is zero, based on approximate computations:
-    >> PossibleZeroQ[2^(2 I) - 2^(-2 I) - 2 I Sin[Log[4]]]
-     = True
-
-    >> PossibleZeroQ[Sqrt[x^2] - x]
-     = False
-    """
-
-    summary_text = "test whether an expression is estimated to be zero"
-    attributes = A_LISTABLE | A_NUMERIC_FUNCTION | A_PROTECTED
-
-    sympy_name = "_iszero"
-
-    def eval(self, expr, evaluation):
-        "%(name)s[expr_]"
-        from sympy.matrices.utilities import _iszero
-
-        sympy_expr = expr.to_sympy()
-        result = _iszero(sympy_expr)
-        if result is None:
-            # try expanding the expression
-            exprexp = Expression(SymbolExpandAll, expr).evaluate(evaluation)
-            exprexp = exprexp.to_sympy()
-            result = _iszero(exprexp)
-        if result is None:
-            # Can't get exact answer, so try approximate equal
-            numeric_val = eval_N(expr, evaluation)
-            if numeric_val and hasattr(numeric_val, "is_approx_zero"):
-                result = numeric_val.is_approx_zero
-            elif not numeric_val.is_numeric(evaluation):
-                return (
-                    SymbolTrue
-                    if Expression(SymbolSimplify, expr).evaluate(evaluation) == Integer0
-                    else SymbolFalse
-                )
-
-        return from_python(result)
 
 
 class Product(IterationFunction, SympyFunction):
