+++ mathics/builtin/matrices/partmatrix.py
@@ -62,42 +62,4 @@ class Diagonal(Builtin):
         return ListExpression(*result)
 
 
-class MatrixQ(Builtin):
-    """
-    <url>
-    :WMA link:
-    https://reference.wolfram.com/language/ref/MatrixQ.html</url>
-
-    <dl>
-      <dt>'MatrixQ[$m$]'
-      <dd>gives 'True' if $m$ is a list of equal-length lists.
-
-      <dt>'MatrixQ[$m$, $f$]'
-      <dd>gives 'True' only if '$f$[$x$]' returns 'True' for when applied to \
-         element $x$ of the matrix $m$.
-    </dl>
-
-    >> MatrixQ[{{1, 3}, {4.0, 3/2}}, NumberQ]
-     = True
-
-    These are not matrices:
-    >> MatrixQ[{{1}, {1, 2}}] (* first row should have length two *)
-     = False
-
-    >> MatrixQ[Array[a, {1, 1, 2}]]
-     = False
-
-    Supply a test function parameter to generalize and specialize:
-    >> MatrixQ[{{1, 2}, {3, 4 + 5}}, Positive]
-     = True
-
-    >> MatrixQ[{{1, 2 I}, {3, 4 + 5}}, Positive]
-     = False
-    """
-
-    rules = {
-        "MatrixQ[expr_]": "ArrayQ[expr, 2]",
-        "MatrixQ[expr_, test_]": "ArrayQ[expr, 2, test]",
-    }
-
-    summary_text = "gives 'True' if the given argument is a list of equal-length lists"
+# TODO: add ArrayRules, Indexed, LowerTriangularize, UpperTriangularize
