+++ mathics/builtin/procedural.py
@@ -18,7 +18,6 @@ environment.
 
 
 from mathics.builtin.base import BinaryOperator, Builtin, IterationFunction
-from mathics.builtin.patterns import match
 from mathics.core.attributes import (
     A_HOLD_ALL,
     A_HOLD_REST,
@@ -35,6 +34,7 @@ from mathics.core.interrupt import (
 )
 from mathics.core.symbols import Symbol, SymbolFalse, SymbolNull, SymbolTrue
 from mathics.core.systemsymbols import SymbolMatchQ
+from mathics.eval.patterns import match
 
 SymbolWhich = Symbol("Which")
