+++ mathics/builtin/tensors.py
@@ -26,7 +26,6 @@ from mathics.core.attributes import A_FLAT, A_ONE_IDENTITY, A_PROTECTED
 from mathics.core.evaluation import Evaluation
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
-from mathics.core.rules import Pattern
 from mathics.core.symbols import Atom, Symbol, SymbolFalse, SymbolTrue
 
 
@@ -103,75 +102,6 @@ class ArrayDepth(Builtin):
     }
 
     summary_text = "the rank of a tensor"
-
-
-class ArrayQ(Builtin):
-    """
-    <url>:WMA: https://reference.wolfram.com/language/ref/ArrayQ.html</url>
-
-    <dl>
-      <dt>'ArrayQ[$expr$]'
-      <dd>tests whether $expr$ is a full array.
-
-      <dt>'ArrayQ[$expr$, $pattern$]'
-      <dd>also tests whether the array depth of $expr$ matches $pattern$.
-
-      <dt>'ArrayQ[$expr$, $pattern$, $test$]'
-      <dd>furthermore tests whether $test$ yields 'True' for all elements of $expr$.
-        'ArrayQ[$expr$]' is equivalent to 'ArrayQ[$expr$, _, True&]'.
-    </dl>
-
-    >> ArrayQ[a]
-     = False
-    >> ArrayQ[{a}]
-     = True
-    >> ArrayQ[{{{a}},{{b,c}}}]
-     = False
-    >> ArrayQ[{{a, b}, {c, d}}, 2, SymbolQ]
-     = True
-    """
-
-    rules = {
-        "ArrayQ[expr_]": "ArrayQ[expr, _, True&]",
-        "ArrayQ[expr_, pattern_]": "ArrayQ[expr, pattern, True&]",
-    }
-
-    summary_text = "test whether an object is a tensor of a given rank"
-
-    def eval(self, expr, pattern, test, evaluation: Evaluation):
-        "ArrayQ[expr_, pattern_, test_]"
-
-        pattern = Pattern.create(pattern)
-
-        dims = [len(expr.get_elements())]  # to ensure an atom is not an array
-
-        def check(level, expr):
-            if not expr.has_form("List", None):
-                test_expr = Expression(test, expr)
-                if test_expr.evaluate(evaluation) != SymbolTrue:
-                    return False
-                level_dim = None
-            else:
-                level_dim = len(expr.elements)
-
-            if len(dims) > level:
-                if dims[level] != level_dim:
-                    return False
-            else:
-                dims.append(level_dim)
-            if level_dim is not None:
-                for element in expr.elements:
-                    if not check(level + 1, element):
-                        return False
-            return True
-
-        if not check(0, expr):
-            return SymbolFalse
-
-        depth = len(dims) - 1  # None doesn't count
-        if not pattern.does_match(Integer(depth), evaluation):
-            return SymbolFalse
-        return SymbolTrue
 
 
 class Dimensions(Builtin):
@@ -576,29 +506,3 @@ class Transpose(Builtin):
                 else:
                     result[col_index].append(item)
         return ListExpression(*[ListExpression(*row) for row in result])
-
-
-# Should be in Elements of Vectors, but we don't have this since other operations
-# are subsumed by Elements of Lists.
-class VectorQ(Builtin):
-    """
-    <url>:WMA link:
-    https://reference.wolfram.com/language/ref/VectorQ.html</url>
-
-    <dl>
-      <dt>'VectorQ[$v$]'
-      <dd>returns 'True' if $v$ is a list of elements which are not themselves lists.
-
-      <dt>'VectorQ[$v$, $f$]'
-      <dd>returns 'True' if $v$ is a vector and '$f$[$x$]' returns 'True' for each element $x$ of $v$.
-    </dl>
-
-    >> VectorQ[{a, b, c}]
-     = True
-    """
-
-    rules = {
-        "VectorQ[expr_]": "ArrayQ[expr, 1]",
-        "VectorQ[expr_, test_]": "ArrayQ[expr, 1, test]",
-    }
-    summary_text = "test whether an object is a vector"
