+++ mathics/builtin/testing_expressions/__init__.py
@@ -1,0 +1,9 @@
+"""
+Testing Expresssions
+
+
+There are a number of functions for testing Expressions.
+
+Functions that "ask a question" have names that end in "Q". \
+They return 'True' for an explicit answer, and 'False' otherwise.
+"""
