+++ mathics/core/systemsymbols.py
@@ -67,6 +67,7 @@ SymbolCosh = Symbol("System`Cosh")
 SymbolCot = Symbol("System`Cot")
 SymbolCoth = Symbol("System`Coth")
 SymbolD = Symbol("System`D")
+SymbolDefault = Symbol("System`Default")
 SymbolDefinition = Symbol("System`Definition")
 SymbolDerivative = Symbol("System`Derivative")
 SymbolDirectedInfinity = Symbol("System`DirectedInfinity")
@@ -78,6 +79,7 @@ SymbolEdgeForm = Symbol("System`EdgeForm")
 SymbolEqual = Symbol("System`Equal")
 SymbolEquivalent = Symbol("System`Equivalent")
 SymbolEulerGamma = Symbol("System`EulerGamma")
+SymbolExactNumberQ = Symbol("System`ExactNumberQ")
 SymbolExpandAll = Symbol("System`ExpandAll")
 SymbolExport = Symbol("System`Export")
 SymbolExportString = Symbol("System`ExportString")
@@ -133,6 +135,7 @@ SymbolMathMLForm = Symbol("System`MathMLForm")
 SymbolMatrixPower = Symbol("System`MatrixPower")
 SymbolMax = Symbol("System`Max")
 SymbolMaxPrecision = Symbol("System`$MaxPrecision")
+SymbolMaxExtraPrecision = Symbol("System`$MaxExtraPrecision")
 SymbolMean = Symbol("System`Mean")
 SymbolMemberQ = Symbol("System`MemberQ")
 SymbolMessageName = Symbol("System`MessageName")
@@ -147,6 +150,7 @@ SymbolNone = Symbol("System`None")
 SymbolNorm = Symbol("System`Norm")
 SymbolNormal = Symbol("System`Normal")
 SymbolNot = Symbol("System`Not")
+SymbolNothing = Symbol("System`Nothing")
 SymbolNumberForm = Symbol("System`NumberForm")
 SymbolNumberQ = Symbol("System`NumberQ")
 SymbolNumericQ = Symbol("System`NumericQ")
